Move canvas classes to new package powp.jobs2d.canvas
