+++ src/main/java/edu/kis/powp/jobs2d/canvas/Canvas.java
@@ -1,4 +1,4 @@
-package edu.kis.powp.jobs2d.command.canvas;
+package edu.kis.powp.jobs2d.canvas;
 
 
 import edu.kis.powp.jobs2d.command.CompoundCommand;
