+++ src/main/java/edu/kis/powp/jobs2d/canvas/CanvasA3.java
@@ -1,4 +1,4 @@
-package edu.kis.powp.jobs2d.command.canvas;
+package edu.kis.powp.jobs2d.canvas;
 
 
 public class CanvasA3 extends RectangularCanvas {
