+++ src/main/java/edu/kis/powp/jobs2d/canvas/CanvasA4.java
@@ -1,4 +1,4 @@
-package edu.kis.powp.jobs2d.command.canvas;
+package edu.kis.powp.jobs2d.canvas;
 
 
 public class CanvasA4 extends RectangularCanvas {
