+++ src/main/java/edu/kis/powp/jobs2d/features/CanvasFeature.java
@@ -1,7 +1,7 @@
 package edu.kis.powp.jobs2d.features;
 
 import edu.kis.powp.appbase.Application;
-import edu.kis.powp.jobs2d.command.canvas.Canvas;
+import edu.kis.powp.jobs2d.canvas.Canvas;
 import edu.kis.powp.jobs2d.drivers.CanvasManager;
 import edu.kis.powp.jobs2d.drivers.SelectCanvasMenuOptionListener;
 import edu.kis.powp.jobs2d.drivers.gui.UpdateCanvasInfoObserver;
