+++ src/test/java/edu/kis/powp/jobs2d/TestJobs2dApp.java
@@ -10,8 +10,8 @@ import edu.kis.legacy.drawer.shape.LineFactory;
 import edu.kis.powp.appbase.Application;
 import edu.kis.powp.jobs2d.command.ImporterFactory;
 import edu.kis.powp.jobs2d.command.JsonCommandImporter;
-import edu.kis.powp.jobs2d.command.canvas.CanvasA3;
-import edu.kis.powp.jobs2d.command.canvas.CanvasA4;
+import edu.kis.powp.jobs2d.canvas.CanvasA3;
+import edu.kis.powp.jobs2d.canvas.CanvasA4;
 import edu.kis.powp.jobs2d.command.gui.CommandManagerWindow;
 import edu.kis.powp.jobs2d.command.gui.CommandManagerWindowCommandChangeObserver;
 import edu.kis.powp.jobs2d.drivers.*;
